Add stack tag level support to event-subscription resource
The event-subscription and modify tags can be tagged with only resource-level tags as of now, stack level tags are not added as the part of this process to the resource.
Stack Level Tags: are tags specified in the AWS console, they should be shared by all resources created inside that stack.
Resource Level Tags: are tags specified in CloudFormation templates.
This commit provides support to add stack level tags to the resource